requirements.txt
blinker==1.8.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.9
filelock==3.16.1
Flask==3.0.3
identify==2.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
ldap3==2.9.1
load-dotenv==0.1.0
MarkupSafe==3.0.2
nodeenv==1.9.1
numpy==2.1.2
pandas==2.2.3
platformdirs==4.3.6
pre_commit==4.0.1
pyasn1==0.6.1
pysmb==1.2.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
schedule==1.2.2
six==1.16.0
tqdm==4.66.6
tzdata==2024.2
virtualenv==20.27.1
Werkzeug==3.0.6
